+++ lisp/checker/check-eask.el
@@ -15,10 +15,10 @@
 
 ;;; Code:
 
-(load (expand-file-name
-       "../_prepare.el"
-       (file-name-directory (nth 1 (member "-scriptload" command-line-args))))
-      nil t)
+(let ((dir (file-name-directory (nth 1 (member "-scriptload" command-line-args)))))
+  (load (expand-file-name "_prepare.el"
+                          (locate-dominating-file dir "_prepare.el"))
+        nil t))
 
 ;; Plain Text
 (defvar eask--checker-log nil)